+++ src/interactions/modals/cakedEventSubmit.ts
@@ -1,15 +1,16 @@
 import { type ModalSubmitInteraction, MessageFlags, ContainerBuilder, TextDisplayBuilder } from 'discord.js'
 import { CAKED_COLOR } from '../../commands/caked'
 import { sep } from '../../utils/cv2'
+import { safeInlineCode, safeMarkdown } from '../../utils/escape'
 
 export async function handleCakedEventSubmit(
   interaction: ModalSubmitInteraction
 ): Promise<void> {
-  const items = interaction.fields.getTextInputValue('items')
-  const people = interaction.fields.getTextInputValue('people')
-  const restrictions = interaction.fields.getTextInputValue('restrictions')
-  const taste = interaction.fields.getTextInputValue('taste')
-  const dateandtime = interaction.fields.getTextInputValue('dateandtime')
+  const items = safeMarkdown(interaction.fields.getTextInputValue('items'))
+  const people = safeInlineCode(interaction.fields.getTextInputValue('people'))
+  const restrictions = safeMarkdown(interaction.fields.getTextInputValue('restrictions'))
+  const taste = safeInlineCode(interaction.fields.getTextInputValue('taste'))
+  const dateandtime = safeInlineCode(interaction.fields.getTextInputValue('dateandtime'))
 
   const container = new ContainerBuilder()
     .setAccentColor(CAKED_COLOR)